Bug 243166 - NullPointerException in java.io.File constructor

+++ bundles/org.eclipse.equinox.p2.directorywatcher/src/org/eclipse/equinox/internal/provisional/p2/directorywatcher/RepositoryListener.java
@@ -14,6 +14,8 @@ import java.io.File;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.*;
+import org.eclipse.core.runtime.IStatus;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.equinox.internal.p2.core.helpers.LogHelper;
 import org.eclipse.equinox.internal.provisional.p2.artifact.repository.*;
 import org.eclipse.equinox.internal.provisional.p2.core.ProvisionException;
@@ -221,7 +223,13 @@ public class RepositoryListener extends DirectoryChangeListener {
 					IInstallableUnit iu = (IInstallableUnit) candidate;
 					if (changes.contains(iu))
 						return true;
-					File iuFile = new File(iu.getProperty(FILE_NAME));
+					String filename = iu.getProperty(FILE_NAME);
+					if (filename == null) {
+						String message = NLS.bind(Messages.filename_missing, "installable unit", iu.getId()); //$NON-NLS-1$
+						LogHelper.log(new Status(IStatus.ERROR, Activator.ID, message, null));
+						return false;
+					}
+					File iuFile = new File(filename);
 					return removedFiles.contains(iuFile);
 				}
 			};
@@ -249,9 +257,15 @@ public class RepositoryListener extends DirectoryChangeListener {
 				IArtifactDescriptor[] descriptors = artifactRepository.getArtifactDescriptors(key);
 				for (int i = 0; i < descriptors.length; i++) {
 					ArtifactDescriptor descriptor = (ArtifactDescriptor) descriptors[i];
-					File artifactFile = new File(descriptor.getRepositoryProperty(FILE_NAME));
-					if (removedFiles.contains(artifactFile))
-						artifactRepository.removeDescriptor(descriptor);
+					String filename = descriptor.getRepositoryProperty(FILE_NAME);
+					if (filename == null) {
+						String message = NLS.bind(Messages.filename_missing, "artifact", descriptor.getArtifactKey()); //$NON-NLS-1$
+						LogHelper.log(new Status(IStatus.ERROR, Activator.ID, message, null));
+					} else {
+						File artifactFile = new File(filename);
+						if (removedFiles.contains(artifactFile))
+							artifactRepository.removeDescriptor(descriptor);
+					}
 				}
 			}
 		}
@@ -268,9 +282,15 @@ public class RepositoryListener extends DirectoryChangeListener {
 			Collector ius = metadataRepository.query(InstallableUnitQuery.ANY, new Collector(), null);
 			for (Iterator it = ius.iterator(); it.hasNext();) {
 				IInstallableUnit iu = (IInstallableUnit) it.next();
-				File iuFile = new File(iu.getProperty(FILE_NAME));
-				Long iuLastModified = new Long(iu.getProperty(FILE_LAST_MODIFIED));
-				currentFiles.put(iuFile, iuLastModified);
+				String filename = iu.getProperty(FILE_NAME);
+				if (filename == null) {
+					String message = NLS.bind(Messages.filename_missing, "installable unit", iu.getId()); //$NON-NLS-1$
+					LogHelper.log(new Status(IStatus.ERROR, Activator.ID, message, null));
+				} else {
+					File iuFile = new File(filename);
+					Long iuLastModified = new Long(iu.getProperty(FILE_LAST_MODIFIED));
+					currentFiles.put(iuFile, iuLastModified);
+				}
 			}
 		}
 		//
